Add files via upload
The project involves the analysis and visualization of Airbnb data using Power BI. The dataset encompasses information about property listings, hosts, prices, room types, availability, reviews, and neighborhood groups. The goal is to provide insights into various aspects of the Airbnb platform.
Analysis of Neighborhood Groups: The report identifies the Neighborhood group with the highest number of listed properties, allowing for an understanding of listing distribution across different areas.

Top-performing Hosts: The report highlights the host who has listed the most properties, recognizing their contributions to the platform.

Success in Popular Neighborhood Groups: The analysis identifies the host with the most properties listed in the neighborhood group with the highest number of listings, showcasing their success in a highly sought-after area.

Average Price Analysis: The report calculates and presents the average price of the listed properties, providing an overview of price ranges within the dataset.

Factors Influencing Prices: The report investigates the factors that contribute to higher prices in specific neighborhood groups, such as proximity to attractions, amenities, or exclusivity of the area.

In conclusion, this Power BI project provides a comprehensive analysis of Airbnb data, allowing users to gain valuable insights into property listings, hosts, prices, room types, availability, reviews, and neighborhood groups. The visualizations and interactive features empower users to explore the data and make informed decisions based on the findings.